+++ packages/alibaba/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @ai-sdk/alibaba
 
+## 2.0.42
+
+### Patch Changes
+
+- Updated dependencies [a4ba394]
+- Updated dependencies [45099da]
+- Updated dependencies [9e1d1b2]
+- Updated dependencies [a495511]
+  - @ai-sdk/provider@4.0.11
+  - @ai-sdk/provider-utils@5.0.37
+
 ## 2.0.41
 
 ### Patch Changes